Use newer OSX
This make sure to use El Capitan and to make perlcritic work on OSX some
tricks are involed. First we need to install both a newer version of
perl and install cpanm ourselves. Then we need to manually create the
~/perl5 directory or the `cpanm --local-lib` part will fail. Finally in
order to _not_ run Linux twice we need to avoid using `perl: 5.18`
toplevel and instead use a tweaked build matrix. Now we're all good.