Add setting to allow disabling of unicode ruler
This commit adds a single setting to allow users
to disable the use of the unicode em-dash that is being used
as a horizonal ruler character for marking the changed filename.

It defaults to the current bahaviour of using em-dash.

Setting `diff-so-fancy.useUnicodeRuler` to `false` will use a simple
dash (minus sign) instead.