documents pattern pre-load tip only for diffs
Closes #173.

The pattern match errors if no match, which almost certainly
happen if it's a log being paged; so we only want to do this if
it really is a diff.

@rocketraman points out that this misses `git show` diffs, but
that case cannot be covered without introducing the same error
when used for blobs, trees, and tags - which `show` may also show.