disables d-s-f when redirecting output to a file
This commit aims to better support standard patch files, by
disabling colours and header removal when the stdout file
descriptor is opened on a terminal.

The assumption is that if this is _not_ the case, it is most
likely that the diff output is being redirected to a .patch file,
for emailing or some other medium in which fancified output may
be confusing, and not compatible with standard tools.

We do this by internalising `less`, providing the bare minimum
`--RAW-CONTROL-CHARS` option required for diff-so-fancy to operate
as expected, and appending any user-supplied options.

This *IS* a breaking change. Configurations that are not updated
will perform a regular non-fancy diff. This is unavoidable since
old configurations piped to `less` or `vim` et al; which we
cannot distinguish from redirection to file.

This commit closes #181.