supports deprecated configurations
This commit fixes the implementation of disabling d-s-f when
redirecting to a file so as not to revert to standard Unified
Format when piping to a pager - as almost all users will be doing
with the current version.

The reason for not doing this initially was to enable piping diff
output to other tools expecting the standard format before ending
up at a `.patch`.

This means that the feature works *ONLY* if nothing comes between
diff-so-fancy and the redirection - viz.
```shell
git diff > foo.patch # works
git diff | sed 's/SECRET=".*"/SECRET=""/' > foo.patch # broken
```

less performs a similar check which is why it's inclusion is safe
before piping to another instance, or e.g. vim. we can keep it
there; deprecate piping to a pager, and revert this commit in a
later major version.