Don't include ansi color codes in file1 and file2
when processing renamed files. Use the same `(.+?)(\e|\t|$)` regexp used
in other similar parts of the code. The motivation behind this change is
to be able to process the file1 and file2 strings (e.g. get their
length) without the interference of the extra ansi color codes. Also,
the resulting d-s-f output is smaller with this change (34 bytes less
for each "renamed: X to Y" line).