color: parse *all* color specs via git itself
git will parse and return the fallback color when given an empty config
name, e.g. `git config --get-color '' 'blue'`. This is documented in the
git config man pages.

This means *none* of that color spec parsing needs to be reproduced
by diff-so-fancy. This is much simpler, and also means that the color
specs can be consistently represented in the git config, avoiding
mismatches between the accepted colors and or modifiers.