fix: over-permissive regex for added/deleted file meta header
The previous pattern uselessly included `.*`, unlike all the other
patterns, and so would erroneously match on actual diff content lines.

Noticed this because the line itself was filtered out of diffs in this
codebase.

This also tightens the pattern since the file mode will always be six
octal (0-7) chars; and in the git codebase the format string is:
  "%s%snew file mode %06o%s\n"