test: use completely isolated, temporary GIT_CONFIG
This writes a temp git config file during setup/teardown, and exports
GIT_CONFIG=… to that and sets GIT_CONFIG_NOSYSTEM=1. This ensures that
any git invocations _only_ use that configuration.

Note that none of the tests currently test anything other than the
default config; however, if they do, the generation of this test config
should be moved into `setup()` rather than `setup_file()` so that it can
be modified and restored between tests.