Re-enable less termcap initialization
This commit removes usage of the the -X option from diff-so-fancy, which
prevented less termcap initialization. From the manpage:

> ```
> -X or --no-init
>     Disables sending the termcap initialization and deinitialization
>     strings to the terminal. This is sometimes desirable if the
>     deinitialization string does something unnecessary, like clearing
>     the screen.
> ```

The default behavior of `less` is to signal termcap initialization. This
behavior provides a number of nice features, including that the output
disappears when `less` is closed, and that mouse scrolling of the less
viewport works automatically on terminals such as gnome-terminal.

I believe that this is a sensible default, and it ensures that
diff-so-fancy works out of the box with fetaures users have come to
expect of `less`.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>