fix: ignore negative color attributes with tests
The colors in the git config may have attributes and negative versions
of them (no/no- prefix).  The negative versions will no longer cause a
false positive checking for the attribute.

A new test file is specifically for testing the settings of the ansi
codes and the possible settings in the git config.

Signed-off-by: Chris. Webster <chris@webstech.net>