fix and maint: match most git supported color settings (#453)
* maint: set ansi colors in common routine

Setting foreground and background colors is now using the same code.

Signed-off-by: Chris. Webster <chris@webstech.net>

* fix: ignore negative color attributes with tests

The colors in the git config may have attributes and negative versions
of them (no/no- prefix).  The negative versions will no longer cause a
false positive checking for the attribute.

A new test file is specifically for testing the settings of the ansi
codes and the possible settings in the git config.

Signed-off-by: Chris. Webster <chris@webstech.net>

* fix: support git config colors  normal and default

Normal is a placeholder and will be ignored.  Default will generate
ansi codes for the default foreground or background color.

Git validates the config so a bright prefix on these values would not be
allowed.  They are added to the hash to keep the code simple.

Signed-off-by: Chris. Webster <chris@webstech.net>

* fix: support #rgb colors in git config

Git supports diff colors in #rrggbb notation.   diff-so-fancy now
supports them as well.

Signed-off-by: Chris. Webster <chris@webstech.net>

---------

Signed-off-by: Chris. Webster <chris@webstech.net>